% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstraps.R
\name{bootstraps}
\alias{bootstraps}
\title{Draw Bootstraps}
\usage{
bootstraps(df, sample.percent, bootstraps)
}
\arguments{
\item{df}{A dataframe}

\item{sample.percent}{percent of the dataframe to sample}

\item{bootstraps}{number of bootstraps to return}
}
\value{
Vector containing N bootstraps
}
\description{
Draw Bootstraps
}
\examples{
bootstraps(iris,.05,10)
}

