% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/changeclass.r
\name{changeclass}
\alias{changeclass}
\title{Correct column classes.}
\usage{
changeclass(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
Processed dataframe that uses heurisitics to set columns to their apprpriate classification
}
\description{
Correct column classes.
}
\examples{
changeclass(mtcars)
}

