% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_engineer.R
\name{date_engineer}
\alias{date_engineer}
\title{Feature Engineering Dates}
\usage{
date_engineer(df, col, holiday)
}
\arguments{
\item{df}{A dfframe}

\item{col}{is numeric index value of date column to be engineered}

\item{holiday}{True or False if ouput should return date values}
}
\value{
dataframe with date values engineered
}

