% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummycode.r
\name{dummyCode}
\alias{dummyCode}
\title{Create dummy columns.}
\usage{
dummyCode(df, DummyColumns, dropcols)
}
\arguments{
\item{df}{A dataframe}

\item{DummyColumns}{a collection or single value that will be converted to dummy columns. This should be a character value.}

\item{Dropcols}{True or False parameter indicating whether to keep or drop the columns that were dummy coded.}
}
\value{
Processed dataframe with new dummy columns
}
\description{
Create dummy columns.
}
\examples{
dummyCode(mtcars, c('vs','gear','carb'),F)
}

