% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(df, col, mean, median, linear)
}
\arguments{
\item{df}{is the dataframe containing missing values}

\item{col}{is the index of the dataframe containing missing values}

\item{mean}{is a True or False parameter}

\item{median}{is a True or False parameter}

\item{linear}{is a True or False parameter}
}
\value{
the RMSE
}
\description{
This function uses heuristics to decide what is the most appropriate method for value imputation.
}

