% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopDisparteFiles.r
\name{LoopDisparteFiles}
\alias{LoopDisparteFiles}
\title{Calculates Regression Metrics}
\usage{
LoopDisparteFiles(FileDir, Seperator, IsHeader, quote)
}
\arguments{
\item{FileDir}{is your regression model LM or MLR}

\item{Seperator}{is the data used to train your model}

\item{IsHeader}{is the value your model is attempting to predict}

\item{quote}{number of features used to in the model}
}
\description{
Function to calculate regression metrics
}

