% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{MAE}
\alias{MAE}
\title{Mean absolute Error}
\usage{
MAE(y, y_pred)
}
\arguments{
\item{y}{is the value of your actual}

\item{y_pred}{is the value that your model predicted}
}
\value{
the MAE
}
\description{
Function to calculate the mean absolute error
}

