% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{is your inpute value}
}
\value{
A single value containing the mode of input x
}
\description{
Computes the mode of a vector, list, or data frame column vector
}

