% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{Mean Square Error}
\usage{
MSE(y, y_pred)
}
\arguments{
\item{y}{is the value of your actual}

\item{y_pred}{is the value that your model predicted}
}
\value{
the MSE
}
\description{
Function to calculate the mean square error
}

