% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionMetrics.R
\name{regression_metrics}
\alias{regression_metrics}
\title{Calculates Regression Metrics}
\usage{
regression_metrics(model, df, trained_Y, Number_of_predictors)
}
\arguments{
\item{model}{is your regression model LM or MLR}

\item{df}{is the data used to train your model}

\item{trained_Y}{is the value your model is attempting to predict}

\item{Number_of_predictors}{number of features used to in the model}
}
\value{
dataframe of regression metrics
}
\description{
Function to calculate regression metrics
}

