% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Calculate Root Mean Square Error}
\usage{
RMSE(y, y_pred)
}
\arguments{
\item{y}{is the value of your actual}

\item{y_pred}{is the value that your model predicted}
}
\value{
the RMSE
}
\description{
Function to calculate the root mean square
}

