% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Missing_Vars.R
\name{Static_Missing_Vars}
\alias{Static_Missing_Vars}
\title{Drop static columns}
\usage{
Static_Missing_Vars(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
Processed dataframe that removes static or columns that are close to static.
}
\description{
Drop static columns
}
\examples{
Static_Missing_Vars(mtcars)
}

