% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim.r
\name{trim}
\alias{trim}
\title{Remove redundant white space.}
\usage{
trim(x)
}
\arguments{
\item{x}{is the value to be trimmed}
}
\value{
a value without any extra whitespace
}
\description{
Remove whitespace from a character or vector. Note that this will always return a character.
}
\examples{
trim('This value      has    an unnecessary amount    of whitespace    ')

## Performing the trim function on a vector
iris$Species <- paste(iris$Species,'    Species',sep='')
head(iris)
## Correct the whitespacing
iris$Species <- trim(iris$Species)
## View our corrected dataframe
head(iris)
rm(iris)
}

