% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstraps.R
\name{bootstraps}
\alias{bootstraps}
\title{bootstraps}
\usage{
bootstraps(df, sample.percent, samples)
}
\arguments{
\item{df}{A dataframe}

\item{sample.percent}{percent of the dataframe to sample}

\item{samples}{number of bootstraps to return}
}
\value{
Vector containing N bootstraps
}
\description{
bootstraps
}
\examples{
bootstraps(iris,.05,10)
}

